Discord_bot/requirements.txt
py-cord==2.3.2
python-dotenv==0.21.1
httpx==0.28.1

